Phase effects optimization in QUnit
