+++ include/qunit.hpp
@@ -85,7 +85,6 @@ protected:
     std::vector<QEngineShard> shards;
     complex phaseFactor;
     bool doNormalize;
-    bool randGlobalPhase;
     bool useHostRam;
     bool useRDRAND;
     bool isSparse;
